fix: resolve CRC validation overflow and add syncRead retry logic

+++ lib/Dynamixel_ll/src/Dynamixel_ll.h
@@ -596,7 +596,7 @@ private:
      * @param length Number of bytes in the data block.
      * @return uint16_t The computed 16-bit CRC value.
      */
-    uint16_t calculateCRC(const uint8_t *data, uint8_t length);
+    uint16_t calculateCRC(const uint8_t *data, uint16_t length);
 
     /**
      * @brief Writes a value to a register.
